requirements.txt
google-generativeai==0.7.2
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.2
oauth2client==4.1.3
python-telegram-bot==21.3
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.2
aiohttp==3.9.5
asyncio==3.4.3
rich==13.7.1